requirements.txt
# AI 이미지 분류기 웹 애플리케이션 - Heroku 배포용

# 웹 프레임워크
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0

# AI/ML 라이브러리 (Heroku 메모리 제한 고려)
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
datasets==2.12.0
accelerate==0.20.3
evaluate==0.4.0

# 이미지 처리
Pillow==10.0.0

# 데이터 처리
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# 불균형 데이터 처리
imbalanced-learn==0.11.0

# 진행률 표시
tqdm==4.65.0

# 유틸리티
python-dotenv==1.0.0

# 로깅
logging==0.4.9.6
